refactor(react-app-env): generalised react-app-env.d.ts files for MFEs
closes #107
